try to structure the use cases
I tried to structure the use cases and initiated some requirements. This is of course a draft version that needs further work. I tried not to remove text, and instead just comment the text that I was not sure is needed. I might have cut a bit too much as I realised some use cases in the direct section were missing. For clarification I am not trying to remove a use case. 

There are still quite a lot of work needed, this is a very early version of the text, but I prefer to  get early discussion to understand if the text is moving in the right direction. 

Here are some notes on the reasoning of the structure. I also leave some comments in the md text <!-- -->.  Please consider the comments below as random comments. I am not entirely convinced these need to be read if the diff can be read by itself.

Structure of the document:

section 3: Discovery of associated resolvers
This section assumes that the client already has established a relation with a resolver that is not encrypted. 
From that starting point, it need to discover the other alternatives.

The primary resolver may be obtained through different ways which provides different level of trust in the relation. 
* DHCP
* configuration
* VPN


The primary resolver may be DoE aware or unaware

The primary may be with local / global IP address

The primary may have specific FQDN scope

In most cases though the scope is defined as: global scope + eventually a private scope and the use is selecting/provided/configured with the one that includes the private scope. 

scope is related to the discovery as the primary resolver may have a specific scope and discovery needs to be enabled with these different scopes. On the other hand, scope probably needs more specification for the selection. I would see such requirements as part of the discovery requirements not as a way to collect appropriated information related to the resolvers but rather as appropriated information that can be provided as input to the selection process. 

In addition to the need be able to perform discover over various types of resolvers that have a limited scope, it also seems that scope an information that we may need to define better in order to provide a more accurate resolver selection. These are more related to the discovered information rather than the discovery protocol but it seems that requirements related to the discovered information remain in scope of this work.

My reading of it is that we are trying to refine this "scope" characteristic that is essential for an appropriated selection and that we are looking at typically the following distinctions:
* explicit list of scopes that may not contain the global scope.
* global scope only (default)

scope needs to be defined, but we may need a representation for no scope as well as global scope. 
While IP addresses and FQDNs might be useful to designate scope, the combination of IP FDQN may define completely different scopes. ( home.arpa IPsubnet0 ) and ( home.arpa IPsubnet1) are different scopes. There is also the case where the scope example.com does not includes subdomain.example.com, though I am not entirely sure. 



While section 3 takes the hypothesis that one starts with one resolver unencrypted resolver to discover the other alternative is a short term necessary step, but that we may also envision a longer term vision where the client is provisioned with an encrypted alternative or with some more abstract parameters that enables the discovery of the full set of resolvers without starting from one specific instance. One motivation is that informations associated to one instances are subject to change over time and a more abstract designation is likely to be able to handle such changes.

While I believe that discovery of alternate resolver can include the case where the primary resolver is encrypted and I do only see the need to create a new section for a discovery input different from a resolver itself.

I would thus opt for the following structure of the document.


section 3: Discovery of associated resolvers
section 4: Direct Discovery of resolvers
section 5: Discovered information


Some other comment on section 3.1


Section 3

I suggest the following additional requirements:

I suggested R1.1 also considered the discovery of resolvers or absence of of these resolvers. 
"associated" is reflexive, symmetric and transitive, so the discovery should apply in my opinion with any instance of the class, and in our case encrypted or unencrypted resolver. 
Maybe we can also turn the existence of the mechanisms into a property the discovery mechanisms ( or protocol) has. 
So here are the changes I propose on R1.1 

OLD:
There must be a mechanism for a client to learn the set of encrypted resolvers that are associated with an unencrypted resolver.

NEW:
Discovery MUST provide a client to learn the set or absence of associated from any instance of resolver.

I see R1.2 as reflecting the fact that the information that initiate the discover may beof local scope information. I think we may not restrict the information to a local IP address but to any local information that may include a local scope name ( home.arpa) for example. 
I am also not quite sure R1.2 has a nit with encrypted resolver. So unless I am reading R1.2 in a wrong way I would write is as:

R1.2: When Discovery is instantiated from an resolver that resolver MAY be encrypted or not and MAY be globally or locally reachable. 



Section 3.1 

I understand that authenticate that the DNS resolver.... is performed during the TLS handshake while DNS answers are coming after the handshake. Maybe it might ease the reading to place in a different order. At least, bullet 2 and 3 may be exchanged. 

The section 3.1 seems exclusively focused on forwarders but I do not find the necessity to distinguish it from a resolver as described in section 4.1.1. In fact I would always expect the announced resolver of a network to be able to resolve local names. Of course if these resolvers belong to different domain they will have a different definition of what local is. Typically, the home network is a subnetwork of my ISP network which is a subnetwork of a larger ISP. Each of these network have a different understanding of local. 

My understanding of these forwarders is that they perform two set of actions:
* cache the response
* split between internal naming and external naming. More specifically, if the request concerns a local scope name (home.arpa or .local) the name is forwarded to the internal (authoritative or recursive server). If the requested name concerns non local scope queries, the query is forwarded to the ISP resolver.  

I m not entirely sure 4.1.2 and 4.1.3 cannot be included into 3.1. 

What I would suggest is to move 4.1.2 and 4.1.3 to section 3.1 and have that discussion of the different use cases in the header of section 3.1 - basically right after the paragraph "Frequently, network...." and change forwarder to local resolver in the title of section 3.1.1 and section 3.1.2.



Section 3.1.1 unencrypted forwarders

It seems to me that the issue is more focused on a resolver that cannot be upgraded rather than that it is encrypted only. In the case of the forwarder, the consequences are not limited to the interface between the client and the forwarder but considering the communication between the client and the resolver with the forwarder in the middle, none of the segments are protected, i.e the communication between the client and the forwarder and the communication between the forwarder and the resolver.